Add the foundation for select clauses
This function is marked as unsafe, as we cannot verify that the types of
the selected columns are actually correct. The safe wrapper for this
will end up being a compile-time macro, which verifies the types of the
selected columns (this is expected to still allow for an arbitrary SQL
string).

For implementing the safe wrapper, the biggest question is how to get
the from clause statically at compile time. Hopefully switching to
associated constants will allow us to grab the from clause of the source
at compile time. The expected implementation will then create a prepared
statement with the query, and get type information from that.

I had attempted to have `FromSql` imply `Queriable`, however I received
an error about conflicting implementations. The written impl was:

```rust
impl<T, QS> Queriable<QS> for T where
    QS: QuerySource,
    T: FromSql<QS::SqlType>,
{
    type Row = Self;

    fn build(row: Self::Row) -> Self {
        row
    }
}
```

However, this led to an error about the impl for `User` conflicting.
Presumably since we're grabbing an associated type from `QuerySource`,
the compiler can't actually verify that there is no implementation of
`FromSql` for `User`? If so, that is a strange limitation.

As a result of this issue, I am using a macro to define the invocation
instead.