Add some basic macros to improve usability
When compiler plugins land on stable (or for users willing to be on
nightly), neither of these macros will be invoked directly. The call to
`table!` can be automatically inferred from the database schema, and the
call to `queriable!` can be replaced with `#[derive(Queriable)]`.

However, manual implementations of those traits is pretty boilerplatey,
and I intend for this to be the API for those only wanting features from
stable Rust.

I've also changed the structure of the table slightly, as I have an idea
of how to represent columns nicely, and this will aid in doing that.
I've also moved the "boilerplate" type definitions for tuples and
primitves into a separate file, as I expect to be adding more there in
the near future.