Add a safe interface for select
By using a trait to represent a column on a table, we are able to write
a safe interface to modify the select clause, with a known type that
does not require compiler plugins or macros.

This will not handle 100% of use cases, so we will continue to have the
unsafe form which takes a string and an assumed type. The safe wrapper
around that will use a compiler plugin to check the types from the
database at compile time.

This currently doesn't allow for selecting multiple columns, which will
be resolved shortly.

The separation of `select_sql_inner` was to allow passing both `&str`
and `String`. I did not change the signature of `select_sql`, as there
is another generic parameter which is likely to be passed in manually
rather than inferred, and I did not want to force consumers to add
`, _>` to every invocation.