Allow `select` to be used on `InnerJoinQuerySource`
This unfortunately requried making that struct public. What I wanted to
write was this:

```rust
impl<A, C, Left, Right> SelectableColumn<A, Left, InnerJoinQuerySource<Left, Right>>
    for C where
    A: NativeSqlType,
    C: Column<A, Left>,
    Left: Table + JoinTo<Right>,
    Right: Table,
{}

impl<A, C, Left, Right> SelectableColumn<A, Right, InnerJoinQuerySource<Left, Right>>
    for C where
    A: NativeSqlType,
    C: Column<A, Right>,
    Left: Table + JoinTo<Right>,
    Right: Table,
{}
```

However, this actually leads to an overlapping impl, since there's no
evidence that `Left` and `Right` are different types, nor is there a way
to prove it. Once specialization lands, I'll simply need to provdide an
impl for the case where they are the same.

It's worth noting that there's no real reason to provent overlapping
impls on marker traits like this, and I'm planning to open an RFC
proposing that we relax the rules in that case.