Add basic support for the PG array type
Arrays in PG are absurdly complicated. It involves some bits of
information that we don't need, and some bits of information I don't
know how to handle.

For multi-dimensional arrays, it should be as simple as separating
`Array<T>` from `Array<Array<T>>`, and recursing until the number of
dimensions is 1. However, I believe this requires specialization to
handle properly. I'll investigate further.

For arrays which contain NULL, we actually only get information about
whether or not it conains null at all, without any way to actually tell
what values are NULL (this matters for types like `VarChar`, where 0
bytes cannot be used to differentiate between null and empty string).

Additionally, it is unclear to me when `lower_bound` would ever change,
and how I am supposed to react to that situation. For now, we'll just
assert that it is our expected value.