Find take 2
This implements basic usage of `find`, both for standard and
non-standard primary keys. This code feels much better than what we had
before (mostly), and doesn't leak any type constraints that we don't
control. Of course, the implementation of `query_sql_params` is a
complete joke that only handles this exact use case. I'm going to hold
off on cleaning it up until we get to an implementation of `where`,
which will change things up considerably.

Ultimatley we'll need to move the determination of number of params and
length of each one into a trait of some kind, and handle tuples. When we
get to multiple params, I'm definitely going to want to represent them
as a tuple, so that we can statically verify that the correct values are
being supplied for the given types. We can determine these for arbitrary
SQL with a compiler plugin.

I think we'll probably have a separate trait similar to `FromSqlRow`.
I'm wondering if it makes more sense to change the type of `ToSql` to
just return a `Vec<u8>` instead. The benefit of this is that the
multi-param trait can just return a `Vec<Vec<u8>>`, which should answer
all the questions needed. It's unclear whether we can implement this and
support tuple of tuples without specialization.

One other unresolved question is whether an option is the right type for
`find`, or if we should have a `RecordNotFound` error type.