Use compile-fail to ensure the correctness of our type structure
This replaces all "this should fail to compile" comments with an actual
test that they fail to compile. I've attempted to specify as much as I
think is relevant about the error messages, without too tightly coupling
to the exact wording that rustc gives when possible.

I might need to be more loose in the future. It's a shame there's no way
to specify a regex here, because most of the time what I care about is
that it includes the name of a trait that isn't implemented, and the
right error code.