Add support for "left outer joins"
At this point, we are officially separating the concept of a join from
an association, and I am no longer thinking about associations just yet.
We'll figure out what they look like later, this will sit underneath. We
still need to figure out what selecting from more than two tables looks
like, as well.

Implementing this revealed a fatal flaw in our infrastructure thus far.
The type of a column is not concrete, and changes based on the query. In
particular, a outer join can change the type of a column to be nullable.

As such, we end up with the following strucutre:

- The type of a join is (Left, Nullable<Right>)
- The type of a column from the right side when selected from a join is
  Nullable<Original> (note: We still need to figure out if we can avoid
  having `Option<Option<T>>` here)
- A column of `(Nullable<A>, Nullable<B>)` can be treated as
  `Nullable<(A, B)>` instead

These rules then allow fairly ergonomic working with these columns in
various contstructs. Selecting a single column gives you a nullable
value. Any sort of data structure, including tuples, can be treated as
optional instead of its fields being optional.

For a time during implementation, I ended up having the additional
parameter on `SelectableColumn` be an associated type, which I really
wanted to avoid due to the potential changes in marker trait overlap
coming if https://github.com/rust-lang/rfcs/pull/1268 lands. We managed
to have it be fairly painless to make it an input type, however.

We needed to re-introduce `next_is_null`, to implement the nullable
tuple generic. I opted to have it take the number of columns to check,
as this is the best way I could find to check if all of the values are
null, without adding a method on `FromSql` to ask if it's able to handle
nulls, or anything like that.

Once again, this caused some werid changes in the compiler error
messages. Sometimes we're now getting the same error an extra time, or
one time less, or it's asking for a different trait than it was before.
I need to follow up on this.

As I'm writing this, I'm realizing that I'm probably missing a few
tests, which I need to go backfill now.