Minor formatting changes
If you're git blaming this line, I'll save you a click:
[reblame here](https://github.com/sgrif/yaqb/blame/fa854bd1c75a95c35c745f7e294606882fabe293/src/query_source/mod.rs#L73)
[previous change](https://github.com/sgrif/yaqb/commit/fa854bd1c75a95c35c745f7e294606882fabe293)

Sorry for the git churn commit!

The length of the `SelectableColumn` line was making my eyes hurt.
Additionally, we're seeing a bit of churn on this type definition, and
the new formatting should improve the blamability (even though I just
fucked it up)