Make all macros useable from other crates
1bf0c2271b48d6fdfb80efe71292a7baebd28698 has more detailed reasoning.
This workaround will eventually just go away as it can be defined
generically once the rules for overlap are loosened either by
specialization (allowing for lattice) or rust-lang/rfcs#1268. In order
to use joins, this macro will need to be called and passed every column
for the table on each side.