Add `insert_without_return`
Turns out `insert` as it exists today is really annoying to use if you
don't actually want to get a record back out of it, which makes it
particularly annoying in tests.

Once again I looked at `insert` and thought to myself "why does this
`Vec<U>` need to be a vec and not a slice?! As a reminder since this is
now blame tied to another case like that, as a reminder: `values()`
needs to return a tuple of all of the data, and we cannot return
borrowed content inside of the tuple since it also needs to be a
concrete type. This makes it impossible to implement generically for
structs with borrowed content.