Add support for `COUNT(*)`
The resulting API is more consise, and will result in a more efficient
query than what we had before.

I considered re-using the `Count` struct, and having some weird type
that represented `*` in the context that a value isn't being used, but I
don't think that concept exists outside of `COUNT`, and would require
the same amount of code duplication.