Remove unnecessay type constraint
The type `SelectableExpression` never makes use of the `QuerySource`
type constraint, and we can remove it. Aside from letting us delete a
bit of code, this will enable an idea I'm toying with where we separate
the tables available on the query from the source itself, as I try to
reduce the number of things that need to implement `QuerySource`.