`AsBindParam` doesn't need to be generic over the SQL type
I believe I did this originally so I could just implement it in the
primitives macro for the types module. This was lazy and is causing me
problems as I try to move some code into a context where the native sql
type can't be inferred.

We can probably clean this up once specialization lands by having a
blanket impl over `T: ToSql<ST>` -- although we'll end up with a problem
where `ST` isn't constrained, so that might not work. We could also do
that without specialization by having `Option` implement `ValuesToSql`
instead of `ToSql`, as I'm fairly certain it's the only type that's a
problem there.