Add a helper function for transactions used from tests
While porting crates.io to use this library, we noticed that we had a
bug where we were committing a transaction from the tests, making
subsequent runs fail. When we changed the transaction to always fail,
the type inferrence could no longer determine the "success" type,
requiring invoking it as `transaction::<(), _, _>`. This is a common
enough case to warrant a special function for it, which always rolls
back the transaction.