Ensure `Copy` is derived for all query sources where possible
Since we have a lot of methods like `filter` which need to be able to
consume `self`, we want to make sure that we return types which are
`Copy` whenever possible for ergonomics. Ultimately this ends up being
completely meaningless for many cases, since the concrete types often
end up with a size of 0.