Allow both columns and values to be passed to `eq`
I would have loved to have used `Into` from the standard library to
implement this, but an implementation of that trait which would work for
this case would require HKT. As it stands today, we would need to have
the actual expression it gets converted into be a generic type passed
to the `eq` function, which would mean that type inference would fail to
figure out what to do.

This helps to make the function slightly more generic, and lays the
ground work for what I think might be the properly generic implementation
I'd ultimately like to end up with. Given the following table:

```rust
table! {
    users {
        name -> VarChar,
        nickname -> VarChar,
        nullable_nickname -> Nullable<VarChar>,
    }
}
```

I would like for all of the following to compile:

```rust
name.eq("Sean")
name.eq(nickname)
name.eq(nullable_nickname)
nullable_nickname.eq("Sean")
nullable_nickname.eq(Some("Sean"))
```

I'm actually torn on whether to even support the fifth case, as I've not
yet decided on whether I want `WHERE name IS NULL` to be represented as
`name.eq(None)` or `name.is_null()`. As such, I've opted not to support
any case which involves passing options yet (which actually means it's
impossible to query against a nullable column at the moment. We need to
fix that).

I'm not sure if we'll ever be able to support that case properly, as
it's not just a problem of overlapping impls, but actually a problem of
infinite recursion in possible type derivation that prevents passing a
value to a comparison operator for a nullable column.